+++ pkg/proxy/cli_test.go
@@ -16,7 +16,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package main
+package proxy
 
 import (
 	"testing"
@@ -27,7 +27,7 @@ import (
 )
 
 func TestNewOauthProxyApp(t *testing.T) {
-	a := newOauthProxyApp()
+	a := NewOauthProxyApp()
 	assert.NotNil(t, a)
 }
